+++ crates/miden-protocol/src/asset/mod.rs
@@ -21,6 +21,12 @@ pub use nonfungible::{NonFungibleAsset, NonFungibleAssetDetails};
 mod token_symbol;
 pub use token_symbol::TokenSymbol;
 
+mod asset_callbacks;
+pub use asset_callbacks::AssetCallbacks;
+
+mod asset_callbacks_flag;
+pub use asset_callbacks_flag::AssetCallbackFlag;
+
 mod vault;
 pub use vault::{AssetId, AssetVault, AssetVaultKey, AssetWitness, PartialVault};
 
@@ -118,16 +124,9 @@ impl Asset {
 
     /// Returns true if this asset is the same as the specified asset.
     ///
-    /// Two assets are defined to be the same if:
-    /// - For fungible assets, if they were issued by the same faucet.
-    /// - For non-fungible assets, if the assets are identical.
+    /// Two assets are defined to be the same if their vault keys match.
     pub fn is_same(&self, other: &Self) -> bool {
-        use Asset::*;
-        match (self, other) {
-            (Fungible(l), Fungible(r)) => l.is_from_same_faucet(r),
-            (NonFungible(l), NonFungible(r)) => l == r,
-            _ => false,
-        }
+        self.vault_key() == other.vault_key()
     }
 
     /// Returns true if this asset is a fungible asset.